pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "term-coder"
version = "0.1.0"
description = "Terminal coding assistant with repo-aware context"
readme = { text = "Term Coder CLI.", content-type = "text/markdown" }
requires-python = ">=3.10"
authors = [{ name = "Term Coder" }]
license = { text = "MIT" }
dependencies = [
  "typer[all]>=0.12.0",
  "rich>=13.7.0",
  "pyyaml>=6.0.1",
  "gitpython>=3.1.43",
  "tiktoken>=0.7.0; python_version >= '3.10'",
  "sentence-transformers>=3.0.0; python_version >= '3.10'",
]

[project.scripts]
tc = "term_coder.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["term_coder*"]
